requirements.txt
discord.py
python-dotenv
openai
PyNaCl
